auth-dev/requirements.txt
--trusted-host gitlab.renewbuy.in
--extra-index-url https://__token__:8zKkkbRtGyJ1LdxxUshA@gitlab.renewbuy.in/api/v4/projects/103/packages/pypi/simple
rb-utils==2023.4.6.10.33.37
aiofiles==0.8.0
aiosqlite==0.17.0
aiokafka==0.7.2
alembic==1.8.0
anyio==3.6.1
asgiref==3.5.2
asyncpg==0.25.0
bcrypt==4.0.0
black==22.6.0
certifi==2022.6.15
cffi==1.15.1
click==8.1.3
colorama==0.4.5
cryptography==38.0.1
databases==0.6.0
dnspython==2.2.1
docstring-parser==0.12
email-validator==1.2.1
fastapi==0.78.0
FastAPI-SQLAlchemy==0.2.1
fastapi-users==10.1.5
fastapi-users-db-sqlalchemy==4.0.3
greenlet==1.1.2
h11==0.12.0
h2==4.1.0
hpack==4.0.0
httpcore==0.15.0
httpx==0.23.0
hypercorn==0.13.2
hyperframe==6.0.1
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.8.0
inflection==0.5.1
Jinja2==3.1.2
makefun==1.14.0
Mako==1.2.1
MarkupSafe==2.1.1
mypy-extensions==0.4.3
ormar==0.11.1
passlib==1.7.4
pathspec==0.9.0
piccolo==0.80.1
piccolo-admin==0.21.0
piccolo-api==0.38.0
platformdirs==2.5.2
priority==2.0.0
psycopg2-binary==2.9.3
pycparser==2.21
pydantic==1.9.1
pandas==1.5.2
PyJWT==2.4.0
python-multipart==0.0.5
python-dotenv==0.21.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
sqladmin==0.1.11
SQLAlchemy==1.4.37
starlette==0.19.1
targ==0.3.7
toml==0.10.2
tomli==2.0.1
typing_extensions==4.2.0
uvicorn==0.17.6
wsproto==1.1.0
WTForms==3.0.1
zipp==3.8.0
fastapi-pagination==0.9.3
python-dateutil==2.8.2
